+++ src/dom/insertion.js
@@ -13,73 +13,56 @@ KISSY.add('dom/insertion', function(S, DOM) {
              * Inserts the new node as the previous sibling of the reference node.
              * @return {HTMLElement} The node that was inserted (or null if insert fails)
              */
-            insertBefore: function(newNode, refNode) {
-                if ((newNode = DOM.get(newNode))
-                    && (refNode = DOM.get(refNode))
-                    && refNode[PARENT_NODE]) {
-                    refNode[PARENT_NODE].insertBefore(newNode, refNode);
+            insertBefore: function(newNodes, refNode) {
+                newNodes = DOM.query(newNodes);
+                if (refNode = DOM.get(refNode) && refNode[PARENT_NODE]) {
+                    newNodes.each(function(newNode) {
+                        refNode[PARENT_NODE].insertBefore(newNode, refNode);
+                    });
                 }
-                return newNode;
             },
 
             /**
              * Inserts the new node as the next sibling of the reference node.
              * @return {HTMLElement} The node that was inserted (or null if insert fails)
              */
-            insertAfter: function(newNode, refNode) {
-                if ((newNode = DOM.get(newNode))
-                    && (refNode = DOM.get(refNode))
-                    && refNode[PARENT_NODE]) {
-                    if (refNode[NEXT_SIBLING]) {
+            insertAfter: function(newNodes, refNode) {
+                newNodes = DOM.query(newNodes);
+                if (refNode = DOM.get(refNode) && refNode[PARENT_NODE]) {
+                    newNodes.each(function(newNode) {
                         refNode[PARENT_NODE].insertBefore(newNode, refNode[NEXT_SIBLING]);
-                    } else {
-                        refNode[PARENT_NODE].appendChild(newNode);
-                    }
+                    });
                 }
-                return newNode;
             },
 
             /**
              * Inserts the new node as the last child.
              */
-            append: function(node, parent) {
-                if ((node = DOM.get(node)) && (parent = DOM.get(parent))) {
+            append: function(nodes, parent) {
+                nodes = DOM.query(nodes);
+                if (parent = DOM.get(parent)) {
                     if (parent.appendChild) {
-                        parent.appendChild(node);
+                        nodes.each(function(node) {
+                            parent.appendChild(node);
+                        });
                     }
                 }
-                return node;
-            },
-
-            /**
-             * KISSY Node need
-             */
-            appendTo:function(node, parent) {
-                return DOM.append(node, parent);
             },
 
             /**
              * Inserts the new node as the first child.
              */
-            prepend:function(node, parent) {
-                if ((node = DOM.get(node))
-                    && (parent = DOM.get(parent))) {
-                    if (parent.firstChild) {
-                        DOM.insertBefore(node, parent.firstChild);
-                    } else {
-                        parent.appendChild(node);
-                    }
+            prepend:function(nodes, parent) {
+                nodes = DOM.query(nodes);
+                if (parent = DOM.get(parent)) {
+                    nodes.each(function(node) {
+                        parent.insertBefore(node, parent.firstChild);
+                    });
                 }
-                return node;
-            },
-
-            /**
-             * KISSY Node need
-             */
-            prependTo:function(node, parent) {
-                return DOM.prepend(node, parent);
             }
         });
+    DOM.prependTo = DOM.prepend;
+    DOM.appendTo = DOM.append;
     return DOM;
 }, {
         requires:["dom/base"]