+++ src/dom/selector.js
@@ -5,6 +5,7 @@
 KISSY.add('dom/selector', function(S, DOM, undefined) {
 
     var doc = document,
+        isNodeList = DOM._isNodeList,
         SPACE = ' ',
         ANY = '*',
         REG_ID = /^#[\w-]+$/,
@@ -120,15 +121,6 @@ KISSY.add('dom/selector', function(S, DOM, undefined) {
         return ret;
     }
 
-    // Ref: http://lifesinger.github.com/lab/2010/nodelist.html
-    function isNodeList(o) {
-        // 注1：ie 下，有 window.item, typeof node.item 在 ie 不同版本下，返回值不同
-        // 注2：select 等元素也有 item, 要用 !node.nodeType 排除掉
-        // 注3：通过 namedItem 来判断不可靠
-        // 注4：getElementsByTagName 和 querySelectorAll 返回的集合不同
-        // 注5: 考虑 iframe.contentWindow
-        return o && !o.nodeType && o.item && !o.setTimeout;
-    }
 
     // 调整 context 为合理值
     function tuneContext(context) {